[fc] Adding range based retry config for batch get
