+++ services/venice-router/src/main/java/com/linkedin/venice/router/VeniceRouterConfig.java
@@ -114,6 +114,7 @@ import com.linkedin.venice.router.api.RoutingComputationMode;
 import com.linkedin.venice.router.api.VeniceMultiKeyRoutingStrategy;
 import com.linkedin.venice.router.api.routing.helix.HelixGroupSelectionStrategyEnum;
 import com.linkedin.venice.router.httpclient.StorageNodeClientType;
+import com.linkedin.venice.utils.BatchGetConfigUtils;
 import com.linkedin.venice.utils.LogContext;
 import com.linkedin.venice.utils.RegionUtils;
 import com.linkedin.venice.utils.Time;
@@ -121,7 +122,6 @@ import com.linkedin.venice.utils.Utils;
 import com.linkedin.venice.utils.VeniceProperties;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
 import java.util.concurrent.TimeUnit;
@@ -263,7 +263,7 @@ public class VeniceRouterConfig implements RouterRetryConfig {
       sslToStorageNodes = props.getBoolean(SSL_TO_STORAGE_NODES, false);
       maxReadCapacityCu = props.getLong(MAX_READ_CAPACITY, 100000);
       longTailRetryForSingleGetThresholdMs = props.getInt(ROUTER_LONG_TAIL_RETRY_FOR_SINGLE_GET_THRESHOLD_MS, 15);
-      longTailRetryForBatchGetThresholdMs = parseRetryThresholdForBatchGet(
+      longTailRetryForBatchGetThresholdMs = BatchGetConfigUtils.parseRetryThresholdForBatchGet(
           props.getString(
               ROUTER_LONG_TAIL_RETRY_FOR_BATCH_GET_THRESHOLD_MS,
               "1-5:15,6-20:30,21-150:50,151-500:100,501-:500"));
@@ -732,79 +732,6 @@ public class VeniceRouterConfig implements RouterRetryConfig {
 
   public String getSystemSchemaClusterName() {
     return systemSchemaClusterName;
-  }
-
-  /**
-   * The expected config format is like the following:
-   * "1-10:20,11-50:50,51-200:80,201-:1000"
-   */
-  public static TreeMap<Integer, Integer> parseRetryThresholdForBatchGet(String retryThresholdStr) {
-    final String retryThresholdListSeparator = ",\\s*";
-    final String retryThresholdSeparator = ":\\s*";
-    final String keyRangeSeparator = "-\\s*";
-    String[] retryThresholds = retryThresholdStr.split(retryThresholdListSeparator);
-    List<String> retryThresholdList = Arrays.asList(retryThresholds);
-    // Sort by the lower bound of the key ranges.
-    retryThresholdList.sort((range1, range2) -> {
-      // Find the lower bound of the key ranges
-      String[] keyRange1 = range1.split(keyRangeSeparator);
-      String[] keyRange2 = range2.split(keyRangeSeparator);
-      if (keyRange1.length != 2) {
-        throw new VeniceException(
-            "Invalid single retry threshold config: " + range1 + ", which contains two parts separated by '"
-                + keyRangeSeparator + "'");
-      }
-      if (keyRange2.length != 2) {
-        throw new VeniceException(
-            "Invalid single retry threshold config: " + range2 + ", which should contain two parts separated by '"
-                + keyRangeSeparator + "'");
-      }
-      return Integer.parseInt(keyRange1[0]) - Integer.parseInt(keyRange2[0]);
-    });
-    TreeMap<Integer, Integer> retryThresholdMap = new TreeMap<>();
-    // Check whether the key ranges are continuous, and store the mapping if everything is good
-    int previousUpperBound = 0;
-    final int MAX_KEY_COUNT = Integer.MAX_VALUE;
-    for (String singleRetryThreshold: retryThresholdList) {
-      // parse the range and retry threshold
-      String[] singleRetryThresholdParts = singleRetryThreshold.split(retryThresholdSeparator);
-      if (singleRetryThresholdParts.length != 2) {
-        throw new VeniceException(
-            "Invalid single retry threshold config: " + singleRetryThreshold + ", which"
-                + " should contain two parts separated by '" + retryThresholdSeparator + "'");
-      }
-      Integer threshold = Integer.parseInt(singleRetryThresholdParts[1]);
-      String[] keyCountRange = singleRetryThresholdParts[0].split(keyRangeSeparator);
-      int upperBoundKeyCount = MAX_KEY_COUNT;
-      if (keyCountRange.length > 2) {
-        throw new VeniceException(
-            "Invalid single retry threshold config: " + singleRetryThreshold + ", which"
-                + " should contain only lower bound and upper bound of key count range");
-      }
-      int lowerBoundKeyCount = Integer.parseInt(keyCountRange[0]);
-      if (keyCountRange.length == 2) {
-        upperBoundKeyCount = keyCountRange[1].isEmpty() ? MAX_KEY_COUNT : Integer.parseInt(keyCountRange[1]);
-      }
-      if (lowerBoundKeyCount < 0 || upperBoundKeyCount < 0 || lowerBoundKeyCount > upperBoundKeyCount) {
-        throw new VeniceException("Invalid single retry threshold config: " + singleRetryThreshold);
-      }
-      if (lowerBoundKeyCount != previousUpperBound + 1) {
-        throw new VeniceException(
-            "Current retry threshold config: " + retryThresholdStr + " is not continuous according to key count range");
-      }
-      retryThresholdMap.put(lowerBoundKeyCount, threshold);
-      previousUpperBound = upperBoundKeyCount;
-    }
-    if (!retryThresholdMap.containsKey(1)) {
-      throw new VeniceException(
-          "Retry threshold for batch-get: " + retryThresholdStr + " should be setup starting from 1");
-    }
-    if (previousUpperBound != MAX_KEY_COUNT) {
-      throw new VeniceException(
-          " Retry threshold for batch-get: " + retryThresholdStr + " doesn't cover unlimited key count");
-    }
-
-    return retryThresholdMap;
   }
 
   public int getResolveThreads() {
